[Build] Deploy into distinct p2-repositories per buildtools version
and reference the current one in the container composite repository, in
order to make it also a 'latest' repository.

In order to create a new version (with separate repository) and to
therefore retain the previous state in a p2-repo, one just has to
increase the version of the root pom.xml.